메인 함수 집중 분산
